refactor(indexer-alt): pull out `pipeline`
## Descrption

Separate the `pipeline` abstraction from the `handler` abstraction, and
put it into its own module. This is in preparation for:

- Splitting up the batching and committing processes. This will make it
  possible to test committing multiple chunks in parallel, and it may be
  possible to re-use some of this logic in...
- an `in_order::pipeline` to use for pipelines that need to write out
  data strictly in checkpoint order and along checkpoint boundaries.

## Test plan

Existing tests and CI.
