+++ crates/sui-indexer-alt/src/handlers/ev_emit_mod.rs
@@ -23,7 +23,7 @@ impl Handler for EvEmitMod {
 
     type Value = StoredEvEmitMod;
 
-    fn handle(checkpoint: &Arc<CheckpointData>) -> Result<Vec<Self::Value>> {
+    fn process(checkpoint: &Arc<CheckpointData>) -> Result<Vec<Self::Value>> {
         let CheckpointData {
             transactions,
             checkpoint_summary,
