+++ crates/sui-indexer-alt/src/handlers/ev_struct_inst.rs
@@ -23,7 +23,7 @@ impl Handler for EvStructInst {
 
     type Value = StoredEvStructInst;
 
-    fn handle(checkpoint: &Arc<CheckpointData>) -> Result<Vec<Self::Value>> {
+    fn process(checkpoint: &Arc<CheckpointData>) -> Result<Vec<Self::Value>> {
         let CheckpointData {
             transactions,
             checkpoint_summary,
