+++ crates/sui-indexer-alt/src/handlers/kv_checkpoints.rs
@@ -19,7 +19,7 @@ impl Handler for KvCheckpoints {
 
     type Value = StoredCheckpoint;
 
-    fn handle(checkpoint: &Arc<CheckpointData>) -> Result<Vec<Self::Value>> {
+    fn process(checkpoint: &Arc<CheckpointData>) -> Result<Vec<Self::Value>> {
         let sequence_number = checkpoint.checkpoint_summary.sequence_number as i64;
         Ok(vec![StoredCheckpoint {
             sequence_number,
