+++ crates/sui-indexer-alt/src/handlers/kv_transactions.rs
@@ -23,7 +23,7 @@ impl Handler for KvTransactions {
 
     type Value = StoredTransaction;
 
-    fn handle(checkpoint: &Arc<CheckpointData>) -> Result<Vec<Self::Value>> {
+    fn process(checkpoint: &Arc<CheckpointData>) -> Result<Vec<Self::Value>> {
         let CheckpointData {
             transactions,
             checkpoint_summary,
