+++ crates/sui-indexer-alt/src/pipeline/processor.rs
@@ -1,0 +1,119 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+use std::sync::Arc;
+
+use futures::TryStreamExt;
+use mysten_metrics::spawn_monitored_task;
+use sui_types::full_checkpoint_content::CheckpointData;
+use tokio::{sync::mpsc, task::JoinHandle};
+use tokio_stream::{wrappers::ReceiverStream, StreamExt};
+use tokio_util::sync::CancellationToken;
+use tracing::{debug, error, info};
+
+use crate::{handlers::Handler, metrics::IndexerMetrics};
+
+use super::Indexed;
+
+/// The processor task is responsible for taking checkpoint data and breaking it down into rows
+/// ready to commit. It spins up a supervisor that waits on the `rx` channel for checkpoints, and
+/// distributes them among `H::FANOUT` workers.
+///
+/// Each worker processes a checkpoint into rows and sends them on to the committer using the `tx`
+/// channel.
+///
+/// The task will shutdown if the `cancel` token is cancelled, or if any of the workers encounters
+/// an error -- there is no retry logic at this level.
+pub(super) fn processor<H: Handler + 'static>(
+    rx: mpsc::Receiver<Arc<CheckpointData>>,
+    tx: mpsc::Sender<Indexed<H>>,
+    metrics: Arc<IndexerMetrics>,
+    cancel: CancellationToken,
+) -> JoinHandle<()> {
+    /// Internal type used by workers to propagate errors or shutdown signals up to their
+    /// supervisor.
+    #[derive(thiserror::Error, Debug)]
+    enum Break {
+        #[error("Shutdown received")]
+        Cancel,
+
+        #[error(transparent)]
+        Err(#[from] anyhow::Error),
+    }
+
+    spawn_monitored_task!(async move {
+        info!(pipeline = H::NAME, "Starting handler");
+        match ReceiverStream::new(rx)
+            .map(Ok)
+            .try_for_each_concurrent(H::FANOUT, |checkpoint| {
+                let tx = tx.clone();
+                let metrics = metrics.clone();
+                let cancel = cancel.clone();
+                async move {
+                    if cancel.is_cancelled() {
+                        return Err(Break::Cancel);
+                    }
+
+                    metrics
+                        .total_handler_checkpoints_received
+                        .with_label_values(&[H::NAME])
+                        .inc();
+
+                    let guard = metrics
+                        .handler_checkpoint_latency
+                        .with_label_values(&[H::NAME])
+                        .start_timer();
+
+                    let values = H::process(&checkpoint)?;
+                    let elapsed = guard.stop_and_record();
+
+                    let epoch = checkpoint.checkpoint_summary.epoch;
+                    let cp_sequence_number = checkpoint.checkpoint_summary.sequence_number;
+                    let tx_hi = checkpoint.checkpoint_summary.network_total_transactions;
+
+                    debug!(
+                        pipeline = H::NAME,
+                        checkpoint = cp_sequence_number,
+                        elapsed_ms = elapsed * 1000.0,
+                        "Processed checkpoint",
+                    );
+
+                    metrics
+                        .total_handler_checkpoints_processed
+                        .with_label_values(&[H::NAME])
+                        .inc();
+
+                    metrics
+                        .total_handler_rows_created
+                        .with_label_values(&[H::NAME])
+                        .inc_by(values.len() as u64);
+
+                    tx.send(Indexed {
+                        epoch,
+                        cp_sequence_number,
+                        tx_hi,
+                        values,
+                    })
+                    .await
+                    .map_err(|_| Break::Cancel)?;
+
+                    Ok(())
+                }
+            })
+            .await
+        {
+            Ok(()) => {
+                info!(pipeline = H::NAME, "Checkpoints done, stopping handler");
+            }
+
+            Err(Break::Cancel) => {
+                info!(pipeline = H::NAME, "Shutdown received, stopping handler");
+            }
+
+            Err(Break::Err(e)) => {
+                error!(pipeline = H::NAME, "Error from handler: {e}");
+                cancel.cancel();
+            }
+        };
+    })
+}
